+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [0.4.0][] (TBD)
+
+## Changed
+
+- All values are strings by default regardless of how they arrived.
+
+## Added
+
+- Values can now be converted to built-in or custom types.
+
 # [0.3.2][] (2022-08-01)
 
 ## Fixed
@@ -34,6 +44,7 @@
 
 Initial release.
 
+[0.4.0]: https://github.com/AaronLasseigne/unifig/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/AaronLasseigne/unifig/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/AaronLasseigne/unifig/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/AaronLasseigne/unifig/compare/v0.2.0...v0.3.0